% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStepsBenchmark.R
\name{annualBenchmark}
\alias{annualBenchmark}
\title{Bends a time-serie with an annual frequency one}
\usage{
annualBenchmark(
  hfserie,
  annualserie,
  include.differenciation = FALSE,
  include.rho = FALSE,
  set.coeff = numeric(),
  set.const = numeric(),
  start.coeff.calc = start(annualserie)[1],
  end.coeff.calc = end(annualserie)[1],
  start.benchmark = start(annualserie)[1],
  end.benchmark = end.coeff.calc + 1,
  start.domain = start(hfserie),
  end.domain = c(end.benchmark + 2, frequency(hfserie))
)
}
\arguments{
\item{hfserie}{the bended time-serie.}

\item{annualserie}{a time-serie with a frequency of 1.}

\item{include.differenciation}{a boolean of length 1. TRUE if the regression includes a differenciation}

\item{include.rho}{a boolean of length 1. TRUE if the regression includes a rho}

\item{set.coeff}{an optional double of length 1. The coefficient of x in the regression}

\item{set.const}{an optional constant of length 1. The constant in the regression}

\item{start.coeff.calc}{the starting year for the coefficients estimation.
By default set to the starting year of the annual serie.}

\item{end.coeff.calc}{the ending year for the coefficients estimation
By default set to the ending year of the annual serie.}

\item{start.benchmark}{the starting year for the annual values to be taken into account
By default the starting year of the annual serie.}

\item{end.benchmark}{the starting year for the annual values to be taken into account
By default one year more than the end for the coefficients.}

\item{start.domain}{the start of the output serie. If it is before start.benchmark, the residuals are smoothed until
this value. 
By default the start of the high frequency serie}

\item{end.domain}{the end of the output. If it is after end.benchmark, the residuals are smoothed until this value.
By default 2 years after the end of the benchmark}
}
\value{
A time serie.
}
\description{
Bends a time-serie with a time serie of a frequency of 1. It is a wrapper of twoStepsBenchmark.
}
