% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_maxent.R
\name{beta2}
\alias{beta2}
\title{Generalized beta function}
\usage{
beta2(alpha)
}
\arguments{
\item{alpha}{a non-negative numeric vector}
}
\description{
A generalization of the base `beta` function with any number of variables
}
