% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bflSmooth.R
\name{bflSmooth}
\alias{bflSmooth}
\title{Smooth a time serie}
\usage{
bflSmooth(lfserie, nfrequency, weights = NULL)
}
\arguments{
\item{lfserie}{a time-serie to be smoothed}

\item{nfrequency}{the new high frequency. It must be a multiple of the low frequency.}

\item{weights}{NULL or a time-serie of the same size than the expected high-frequency serie.
If the weights are submitted, it is the rate hfserie/weights that is smoothed. By example,
if weights is a volume and lfserie a value, the result is a disaggregated value with a smoothed
price.}
}
\value{
A time serie of frequency nfrequency
}
\description{
bflSmooth smoothes a time-serie into a time serie of a higher frequency that exactly
aggregate into the higher one. The process followed is Boot, Feibes and Lisman
}
\author{
Arnaud Feldmann
}
