% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_dist.R
\name{choose_dist_disagg}
\alias{choose_dist_disagg}
\title{Choose an Appropriate Disaggregated Distribution}
\usage{
choose_dist_disagg(shares, sds)
}
\arguments{
\item{shares}{A numeric vector representing the shares parameter(s) for the distribution.}

\item{sds}{A numeric vector representing the sds parameter(s) for the distribution, can be NULL or NA.}
}
\value{
A character string indicating the type of disaggregated distribution selected based on the inputs,
Possible return values include 'dir1' for Dirichlet 1, 'dir_maxent' for Dirichlet Maximum Entropy, 'gdir' for Generalized Dirichlet,
or NA if the inputs do not match any predefined distribution type.
}
\description{
This function determines the appropriate disaggregated distribution type based on shares and sds parameters.
It considers various scenarios and constraints related to the input parameters to select from specific distribution types.
}
\examples{
# To determine a distribution with known shares values:
dist_type <- choose_dist_disagg(shares = c(1,2,3), sds = NULL)
}
