% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praislm-methods.R
\name{coef.praislm}
\alias{coef.praislm}
\title{Extracting the coefficients of a prais-lm regression}
\usage{
\method{coef}{praislm}(x)
}
\arguments{
\item{x}{a prais-lm regression}
}
\value{
a numeric vector
even if some of them are equal to zero.
}
\description{
Extracts the coefficients of the prais-lm regression.
}
