% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStepsBenchmark-methods.R
\name{coef.twoStepsBenchmark}
\alias{coef.twoStepsBenchmark}
\title{Extracting the coefficients of a twoStepsBenchmark}
\usage{
\method{coef}{twoStepsBenchmark}(x)
}
\arguments{
\item{x}{a twoStepsBenchmark}
}
\value{
a numeric vector
}
\description{
Extracts the coefficients of the twoStepsBenchark regression.
}
