% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{construction}
\alias{construction}
\title{Total GFCF in construction at current prices}
\format{
A time-serie of frequency 1
}
\source{
\url{https://www.insee.fr/fr/statistiques/2015361}
}
\usage{
construction
}
\description{
An annual time-serie containing the french GFCF in construction from 2000
to 2019.
}
\keyword{datasets}
