% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consumption_catering}
\alias{consumption_catering}
\title{Total consumption in accommodation and food services at current prices}
\format{
A time series of frequency 1
}
\source{
\url{https://www.insee.fr/en/statistiques/serie/010548586}
}
\usage{
consumption_catering
}
\description{
An annual time series containing the consumption in accommodation and food services services from 1999
to 2021, as retrieved the 04/07/2022.
}
\keyword{internal}
