% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{default_col_pal}
\alias{default_col_pal}
\title{Default color palette}
\usage{
default_col_pal(object)
}
\description{
The default color palette for the graphics, imported from the package
\pkg{scales}. The chosen palette function depends on the input object.
}
\keyword{internal}
