% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{default_lty_pal}
\alias{default_lty_pal}
\title{Default color palette}
\usage{
default_lty_pal()
}
\description{
The default palette for the graphics, imported from the package
\pkg{scales}.
}
\keyword{internal}
