% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{default_theme_ggplot}
\alias{default_theme_ggplot}
\title{Default ggplot theme}
\usage{
default_theme_ggplot(object, start, end, show.legend, xlab, ylab, mar)
}
\description{
This is the default theme for the ggplot graphics produced with autoplot
}
\keyword{internal}
