% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_maxent.R
\name{dirichlet_entropy}
\alias{dirichlet_entropy}
\title{Dirichlet entropy function}
\usage{
dirichlet_entropy(x, shares)
}
\arguments{
\item{x}{value of the concentration parameter (gamma)}

\item{shares}{the shares (must sum to 1)}
}
\description{
calculates the negative entropy of a dirichlet distribution given `shares` (sum to 1) and a concentration parameter `x`, so that `alpha = shares * x`
from:  https://en.wikipedia.org/wiki/Dirichlet_distribution#Entropy
}
