% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_tsp}
\alias{extend_tsp}
\title{Extend tsp with lf}
\usage{
extend_tsp(tsphf, lffreq)
}
\arguments{
\item{tsphf}{a numeric of length 3, a tsp of high-frequency}

\item{lffreq}{a numeric of length 1, the low frequency}
}
\value{
a numeric of length 3, a tsp of high-frequency.
}
\description{
This window is the smallest that is all around tsphf
that is compatible with the low frequency.
}
\keyword{internal}
