% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_maxent.R
\name{find_gamma_maxent2}
\alias{find_gamma_maxent2}
\title{Like `find_gamma_maxent`: Finds the Gamma value which maximises the entropy of a Dirichlet distribution
with given alphas (=sector shares), using the nloptr optimazation package.}
\usage{
find_gamma_maxent2(
  shares,
  eval_f = eval_f,
  eval_grad_f = dirichlet_entropy_grad,
  x0 = 1,
  x0_n_tries = 100,
  bounds = c(0.001, 300),
  shares_lb = 0,
  local_opts = list(algorithm = "NLOPT_LD_MMA", xtol_rel = 1e-04),
  opts = list(algorithm = "NLOPT_GD_STOGO", xtol_rel = 1e-04, maxeval = 1000, local_opts
    = local_opts, print_level = 0)
)
}
\arguments{
\item{shares_lb}{lower bound of shares. Only values LARGER than that will be considered (set this to e.g. 1E-4) to increase chance of convergence}

\item{opts}{}
}
\description{
BUT: including the first derivative of the Dir Entropy function `dirichlet_entropy_grad`,
and thus is **much** faster than `find_gamma_maxent`.
}
