% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_dist.R
\name{generate_distribution_agg}
\alias{generate_distribution_agg}
\title{Generate Aggregated Distribution Based on Parameters}
\usage{
generate_distribution_agg(mean = NA, sd = NA, min = -Inf, max = Inf)
}
\arguments{
\item{mean}{The mean for the distribution, default is NA.}

\item{sd}{The standard deviation for the distribution, default is NA.}

\item{min}{The lower boundary for the distribution, default is -Inf.}

\item{max}{The upper boundary for the distribution, default is Inf.}
}
\value{
A random sample from the chosen aggregated distribution or a stop message if no suitable distribution is found.
}
\description{
This function generates a random sample from an aggregated distribution determined by specified mean, standard deviation,
and boundary parameters. It uses the 'choose_dist_agg' function to select an appropriate distribution based on the input parameters
and then generates a random sample from the chosen distribution.
}
\examples{
# Generate a distribution with specified mean and sd:
sample <- generate_distribution_agg(mean = 5, sd = 2)
}
