% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSmooth.R
\name{hfserie_extrap}
\alias{hfserie_extrap}
\title{Extrapolation function for the hfserie in a rateSmooth}
\usage{
hfserie_extrap(hfserie, lffreq)
}
\arguments{
\item{hfserie}{a time-serie, the high frequency serie to extrapolate}

\item{lffreq}{a integer of length 1. The low frequency
to twoStepsBenchmark}
}
\value{
a time-serie, the extrapolated hfserie
}
\description{
This function replaces the incomplete low frequency cycles, at the start and the end of the hfserie,
with respectively the first and the last complete cycles.
It may seem very raw, but it's only used for the weights in \code{bflSmooth}, in order to get a high
frequency rate.
}
