% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.R
\name{in_convergence}
\alias{in_convergence}
\title{Producing the out-of-sample coefficients of a prais-lm regression}
\usage{
in_convergence(object, type = "indicators")
}
\arguments{
\item{object}{an object of class \code{"praislm"} or \code{"twoStepsBenchmark"}.}

\item{type}{\code{"indicators"}, \code{"constant"}, \code{"indicator-constant-2d"},
\code{"all"} or any selection of coefficient names. "indicators", the
default, unselects the outliers and the constant. "all" selects every
coefficients including outliers and the constant. "indicator-constant-2d"
is only valid for univariate regressions, selects the indicator and the
constant and maps the plot method plot to an alternate 2d plot.}
}
\value{
a named matrix time series of multiple columns, one for each selected
coefficient
A \code{"tscomparison"} class is added to the object.
}
\description{
The function \code{in_convergence} returns the out-of-sample coefficients from a
\link{praislm} or a \link{twoStepsBenchmark} object. These coefficients are
estimated, for each low-frequency observation, using all the \emph{previous}
observations. Hence, these values can be used to check the convergence of
the coefficients.
}
\details{
The functions \code{plot} and \code{autoplot} can be used on this object to produce
graphics.
}
\examples{
benchmark <- twoStepsBenchmark(turnover,construction,include.rho = TRUE)
plot(in_convergence(benchmark))
plot(in_convergence(benchmark, type = "constant"))
plot(in_convergence(benchmark, type = "indicator-constant-2d"))
plot(in_convergence(benchmark, type = "all"))
}
\seealso{
\link{in_disaggr} \link{in_revisions} \link{in_scatter}
\link{in_sample} \link{plot.tscomparison}
}
