% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.R
\name{in_dicator}
\alias{in_dicator}
\title{Comparing a benchmark with its indicator}
\usage{
in_dicator(object, type = "changes")
}
\arguments{
\item{object}{an object of class twoStepsBenchmark`}

\item{type}{"levels","levels-rebased", "changes" or "contributions". This
defines the type of output.}
}
\value{
a named matrix time-serie of two columns, one for the
response and the other for the predicted value.
A \code{tscomparison} class is added to the object.
}
\description{
The function \code{in_dicator}takes a \link{twoStepsBenchmark} object
as an input. It produces a comparison between the benchmarked time-serie and the
input.
}
\details{
The functions \code{plot} and \code{autoplot} can be used on this object to produce
graphics.
}
\examples{
benchmark <- twoStepsBenchmark(turnover,construction,include.rho = TRUE)
plot(in_dicator(benchmark))
}
\seealso{
in_sample in_revisions in_scatter plot.tscomparison
}
