% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{melt_matrix}
\alias{melt_matrix}
\title{Melt a matrix into a long data.table}
\usage{
melt_matrix(x, remove_zero = TRUE, remove_non_finite = TRUE)
}
\arguments{
\item{x}{A matrix that will be converted into a long format data.table.}
}
\value{
A long format data.table containing the melted matrix with 'x', 'y', and 'value' columns.
}
\description{
This function converts a matrix into a long format data.table. It first transforms the matrix into a data.table,
then melts it into long format, cleaning up and converting factors and removing non-finite or non-positive values.
}
\examples{
# Assuming 'mat' is a matrix:
long_dt <- melt_matrix(mat)
}
