% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStepsBenchmark-methods.R
\name{model.list}
\alias{model.list}
\title{Extracting the Model list from a twoStepsBenchmark}
\usage{
model.list(x)
}
\arguments{
\item{x}{a twoStepsBenchmark}
}
\value{
A list containing hfserie, lfserie, include.rho, include.differenciation, set.coeff, set. const, start.coeff.calc
and end.coeff.calc.
}
\description{
The model list of a twoStepsBenchmark is a list which contains the parameters provided to the Benchmark.
}
