% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{outliers}
\alias{outliers}
\title{Extracting the standard error}
\usage{
outliers(object, as.ts = FALSE)
}
\arguments{
\item{object}{a praislm or twoStepsBenchmark object.}

\item{as.ts}{a boolean of length 1. If \code{TRUE}, the returned
outliers are returned as a time-serie with (dim and colnames).
If \code{FALSE}, the returned outliers is the named list that was
submitted as a function argument.}
}
\value{
a named list or a time-serie, depending of the
argument \code{"as.ts"}.
}
\description{
The function \code{outliers} returns the outliers
from either a \link{praislm} or a \link{twoStepsBenchmark} object.
}
\keyword{internal}
