% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rbeta3}
\alias{rbeta3}
\title{Generate random numbers from the beta distribution. Similar to `base::rbeta`, but:
- vectorized
- parametrised with mean and sd (derived from: https://en.wikipedia.org/wiki/Beta_distribution#Mean_and_variance)
- Ensures that the sum of the samples do not exceed 1.}
\usage{
rbeta3(n, shares, sds, fix = TRUE, max_iter = 1000)
}
\arguments{
\item{n}{number of samples}

\item{shares}{a vector of shares (between 0 and 1, sum(means) <= 1)}

\item{sds}{vector of standard devaitions of the shares.}

\item{fix}{if set to TRUE (default) the SD of all variables for which the
parameter combination of shares and sd is undedefined ((shares * (1-shares)) < sd^2).
If set to FALSE the function will throw an error message if there are undefined
parameter combinations.}

\item{max_iter}{the maximum number of iterations to sample.}
}
\description{
It is similar to a generalised Dirichlet, however the main functional difference
is that the samples do NOT sum to 1 (but are <= 1).
}
\details{
The constraint that the sample sum is <= 1 is currently achieved by sampling
as long until `n` samples exist that satisfy the constraint. Depending on the
parameter combination this can be quite computational intensive.

Note that the beta distribution is only defined if (mean * (1-mean)) < sd^2
}
