% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rdir1}
\alias{rdir1}
\title{Generate uniform random number from a Dirichlet distribution.
Uniform means each alpha is 1:https://en.wikipedia.org/wiki/Dirichlet_distribution#When_each_alpha_is_1}
\usage{
rdir1(N, length)
}
\arguments{
\item{N}{number of samples}

\item{length}{the number of variables}
}
\value{
a matrix with n rows and `length` cols, each containing a single Dirichlet random deviate
}
\description{
Generate uniform random number from a Dirichlet distribution.
Uniform means each alpha is 1:https://en.wikipedia.org/wiki/Dirichlet_distribution#When_each_alpha_is_1
}
