% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rdir2}
\alias{rdir2}
\title{Similar to `rdir` but adding a threshold argument.
For all variables `i` with an alpha below that threshold, the `rate` argument of
the `rgamma` functions is set to `1 / alpha[i]`, and `shape` is set `1`. This ensures less extreme values
for those variables which presumably originate from the know issues of
generating random gamma distributed number with a very small `shape` parameter.}
\usage{
rdir2(n, alpha, gamma, threshold = 0.01)
}
\arguments{
\item{threshold}{}
}
\description{
The reason for setting rate and shape parameters for small alphas like this is the following:
The gamma distribution (similar to the beta distribution), generates well behaved
random numbers only if the variables SD is not greater than the MEAN.
The gamma distribution can also be parametrised by a mean and sd using the
following relation between mean/sd and the shape and rate parameters:
$$
rate = \frac{\mu}{\sigma^2} \\
shape = (\frac{\mu}{\sigma})^2
$$
(derived from: https://en.wikipedia.org/wiki/Gamma_distribution#Mean_and_variance)
}
\details{
See `?rgamma`: "Note that for smallish values of shape (and moderate scale) a large parts of the mass of the Gamma distribution is on values of x so near zero that they will be represented as zero in computer arithmetic. So rgamma may well return values which will be represented as zero."

Note: the current implementation is rather a pragmatic workaround for this
issue. Both the threshold parameter and the formula to set the `rate` parameter
are somehow arbitrarily chosen. However, since only variables with a very small
mean are affected this shouldn't affect overall results too much.
}
