% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rdir_maxent}
\alias{rdir_maxent}
\title{Generate random numbers from Dirichlet distribution. The concenctration parameter
$\gamma$ (or sometimes also noted $\alpha_0$) is estimated based on the
the maximum entropy principle.}
\usage{
rdir_maxent(n, shares, ...)
}
\arguments{
\item{n}{number of samples}

\item{shares}{Vector containing a best-guess (mean) values for the shares. Must sum to 1!}

\item{...}{other parameters passed to `find_gamma_maxent2`}
}
\value{
a matrix with n rows and length(shares) cols, each containting a single Dirichlet random deviate
}
\description{
See: `?find_gamma_maxent2`
}
