% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdisagg.R
\name{rdisagg}
\alias{rdisagg}
\title{Generate random disaggregates}
\usage{
rdisagg(n, mean_0, sd_0 = NULL, min = 0, max = Inf, shares, sds = NULL)
}
\arguments{
\item{n}{sample size}

\item{mean_0}{the best guess of the aggregate}

\item{sd_0}{the standard deviation of the aggregate}

\item{min}{the lower boundary}

\item{max}{the upper boundary}

\item{shares}{best guesses for the shares (need to sum to one unless they contain NA's). See `?rshares`}

\item{sds}{standard deviations of the shares (can contain NA). Set to NULL (default) if none available.}
}
\description{
Creates a random sample of disaggregates based on the information provided.
The aggregate and the shares are sampled independently. The distribution
from which to sample is determined internally based on the information
provided by the user.
}
