% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reView.R
\name{rePort}
\alias{rePort}
\title{Producing a report}
\usage{
rePort(
  object,
  output_file = NULL,
  launch.browser = if (is.null(output_file)) TRUE else FALSE,
  ...
)
}
\arguments{
\item{object}{a twoStepsBenchmark with an univariate hfserie, a reViewOutput,
or a character of length 1 with the path of their RDS file. If a reViewOutput
is chosen, the former new benchmark is taken as the old one.}

\item{output_file}{The file in which the html should be saved. If \code{NULL}
the file is temporary, and opened in a tab of the default browser.}

\item{launch.browser}{\code{TRUE} or \code{FALSE}. If TRUE, the output is opened in the
browser. Defaults to TRUE if output_file is NULL.}

\item{\dots}{other arguments passed to rmarkdown::render}
}
\description{
This function takes an output of the \link{reView} \pkg{shiny} application
and produces an html report with the same outputs than in shiny.
}
\details{
It can also directly take a \link{twoStepsBenchmark} as an input.
}
\seealso{
reView
}
