% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praislm-methods.R
\name{residuals.praislm}
\alias{residuals.praislm}
\title{Extracting the residuals of a prais-lm regression}
\usage{
\method{residuals}{praislm}(x)
}
\arguments{
\item{x}{a prais-lm regression}
}
\value{
a numeric vector
}
\description{
Extracts the residuals of a prais-lm regression.
}
