% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStepsBenchmark-methods.R
\name{residuals.twoStepsBenchmark}
\alias{residuals.twoStepsBenchmark}
\title{Extracting the residuals of a twoStepsBenchmark}
\usage{
\method{residuals}{twoStepsBenchmark}(x)
}
\arguments{
\item{x}{a twoStepsBenchmark}
}
\value{
a numeric vector
}
\description{
Extracts the residuals of the twoStepsBenchark regression.
}
