% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praislm-methods.R
\name{rho.praislm}
\alias{rho.praislm}
\title{Extracting the rho of a prais-lm regression}
\usage{
\method{rho}{praislm}(x)
}
\arguments{
\item{x}{a prais-lm regression}
}
\value{
a numeric vector of length 1
}
\description{
Extracts the rho of a prais-lm regression.
}
