% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praislm-methods.R
\name{rho}
\alias{rho}
\title{Extract autocorrelation parameter}
\usage{
rho(object)
}
\arguments{
\item{object}{a praislm or twoStepsBenchmark object.}
}
\value{
a double of length 1.
}
\description{
The function `rho` returns the autocorrelation parameter
from either a \link{praislm} or a \link{twoStepsBenchmark} object.
If `include.rho` is `FALSE`, rho returns zero.
}
