% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rshares}
\alias{rshares}
\title{Generate random shares based on the information provided.
Samples alyways sum to 1.}
\usage{
rshares(n, shares, sds = NULL, na_action = "fill", max_iter = 1000, ...)
}
\arguments{
\item{n}{sample size}

\item{shares}{vector containing a best-guess (mean) values for the shares. Can contain NA's if no information is available.}

\item{sds}{Vector of same length as `shares` containing the SDs of the shares. Can contain NA's if no information is available.}

\item{na_action}{what to do with NA's in the `shares` argument. "fill" (default) fills them proportional with the difference `1-sum(shares)`. "remove" removes them from the vector.}

\item{max_iter}{see `?rbeta3`}

\item{...}{other argeuments passed to `rbeta3`}
}
\description{
Compared to generalised Dirichlet the function can handle mixed information,
e.g. SDs/means only for some of the shares.
}
\details{
The function uses a nested sampling approach:
1. sample those shares for which both, mean and SD is available (either from generalised Dirichlet `dirg()` if mean and sd are given for ALL shares, or from beta distiributions `rbeta3()` when mean and sd are only given for some shares.)
2. sample reminder from standard dirichlet with gamma fitted to MaxEnt (`rdir_maxent`)
}
