% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{smoothed.rate}
\alias{smoothed.rate}
\title{Extracting the rate of a threeRuleSmooth}
\usage{
smoothed.rate(object)
}
\arguments{
\item{object}{a threeRuleSmooth object.}
}
\description{
The function \code{smoothed.rate} returns the high-frequency rate
from a \link{threeRuleSmooth} object.
}
\examples{
benchmark <- threeRuleSmooth(turnover,construction); smoothed.rate(benchmark)

}
\keyword{internal}
