% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_matrices}
\alias{sum_matrices}
\title{Sum Columns of Matrices with Matching Column Names}
\usage{
sum_matrices(list_of_matrices)
}
\arguments{
\item{list_of_matrices}{A list where each element is a matrix of numeric values.
Each matrix must have the same number of rows.}
}
\value{
A matrix with the same number of rows as the input matrices and as many columns as there are unique column names across all matrices. The values are the sums of corresponding entries for each column name.
}
\description{
Takes a list of matrices and returns a single matrix with the same number
of rows and unique column names across all input matrices.
For each column name present in any of the matrices, it creates a column
in the resulting matrix, where each entry is the sum of entries from all
matrices with the same column name.
}
\examples{
sample_comb <- list(
  matrix(c(1, 2, 3, 4, 5, 6), nrow = 2),
  matrix(c(7, 8, 9, 10, 11, 12), nrow = 2, dimnames = list(NULL, c("V1", "V2", "V3")))
)
sumMatrices(sample_comb)

}
