% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turnover}
\alias{turnover}
\title{Turnover indicator in construction}
\format{
A time-serie of frequency 12
}
\source{
\url{https://www.insee.fr/fr/statistiques/4636032}
}
\usage{
turnover
}
\description{
An mensual time-serie containing the turnover indicator in construction
from january 2000 to may 2020, as retrieved the 12/08/2020.
}
\keyword{internal}
