% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turnover_catering}
\alias{turnover_catering}
\title{Turnover indicator in accommodation and food services}
\format{
A time series of frequency 12
}
\source{
\url{https://www.insee.fr/en/statistiques/serie/010543406}
}
\usage{
turnover_catering
}
\description{
An monthly time series containing the turnover indicator in accommodation and food services
from january 1999 to april 2022, as retrieved the 04/07/2022.
}
\keyword{internal}
