% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStepsBenchmark.R
\name{annualBenchmark}
\alias{annualBenchmark}
\title{Regress and bends a time-serie with a lower frequency one}
\usage{
annualBenchmark(
  hfserie,
  lfserie,
  include.differenciation = FALSE,
  include.rho = FALSE,
  set.coeff = NULL,
  set.const = NULL,
  start.coeff.calc = start(lfserie)[1L],
  end.coeff.calc = end(lfserie)[1L],
  start.benchmark = start(lfserie)[1L],
  end.benchmark = end.coeff.calc + 1,
  start.domain = start(hfserie),
  end.domain = c(end.benchmark + 2, frequency(hfserie))
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it's a special case of the
\link{twoStepsBenchmark} function that has not proved to be of much practical
value.
}
\keyword{internal}
