% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bflSmooth.R
\name{bflSmooth}
\alias{bflSmooth}
\title{Smooth a time serie}
\usage{
bflSmooth(lfserie, nfrequency)
}
\arguments{
\item{lfserie}{a time-serie to be smoothed}

\item{nfrequency}{the new high frequency. It must be a multiple of the low frequency.}
}
\value{
A time serie of frequency nfrequency
}
\description{
bflSmooth smoothes a time-serie into a time serie of a higher frequency that exactly
aggregate into the higher one. The process followed is Boot, Feibes and Lisman
}
\author{
Arnaud Feldmann
}
