% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bflSmooth.R
\name{bflSmooth}
\alias{bflSmooth}
\title{Smooth a time serie}
\usage{
bflSmooth(lfserie, nfrequency, weights = NULL)
}
\arguments{
\item{lfserie}{a time-serie to be smoothed}

\item{nfrequency}{the new high frequency. It must be a multiple of the low frequency.}

\item{weights}{NULL or a time-serie of the same size than the expected high-frequency serie.
The weights permits, by example, to smooth prices or rates relatively to a high-frequency
account. The results minimizes the squares of the variations, with the constraint that,
if multiplied by the weights then aggregated, they are equal to the low-frequency serie,
multiplied by the aggregated weigths.}
}
\value{
A time serie of frequency nfrequency
}
\description{
bflSmooth smoothes a time-serie into a time serie of a higher frequency that exactly
aggregate into the higher one. The process followed is Boot, Feibes and Lisman
}
\author{
Arnaud Feldmann
}
