% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{default_col_pal}
\alias{default_col_pal}
\title{Default color palette}
\usage{
default_col_pal(object)
}
\description{
The default color palette for the graphics, inspired from the package
\pkg{scales} whose scales can also be used as alternatives.
}
\keyword{internal}
