% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{default_lty_pal}
\alias{default_lty_pal}
\title{Default color palette}
\usage{
default_lty_pal(object)
}
\description{
The default color palette for the graphics. The palette for the objects with
another class than \code{"in_scatter"} is, as seen in \code{linetype_pal} from the
package \pkg{scales}, based on a set supplied by Richard Pearson, University
of Manchester.
}
\details{
Palettes from scales can be used as alternatives.
}
\keyword{internal}
