% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4declaration.R
\docType{class}
\name{disaggR-class}
\alias{disaggR-class}
\alias{Ops,disaggR,vector-method}
\alias{Ops,vector,disaggR-method}
\alias{Ops,disaggR,ts-method}
\alias{Ops,ts,disaggR-method}
\alias{Ops,disaggR,disaggR-method}
\title{Virtual Class "disaggR" Class of disaggregations}
\description{
The \code{disaggR} class is a class contained by the actual disaggregation classes
in the disaggR package. It is a “virtual” class.  disaggR classes are
minimalist S4 classes, containing disaggR and list. This allows Ops group generic
double dispatch but most of the package rather uses S3 methods.
}
\seealso{
the functions \link{twoStepsBenchmark} and \link{threeRuleSmooth}.
}
