% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggR.R
\docType{package}
\name{disaggR-package}
\alias{disaggR}
\alias{disaggR-package}
\title{disaggR: Two-Steps Benchmarks Inspired by the French Quarterly National Accounts Method}
\description{
Two-steps benchmarks inspired by the French National Account method as described
in INSEE Methods n 126 - May 2012.
}
\keyword{internal}
