% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggR.R
\docType{package}
\name{disaggR-package}
\alias{disaggR}
\alias{disaggR-package}
\title{Two-Steps Benchmarks for Time Series Disaggregation}
\description{
The twoStepsBenchmark() function and its wrappers allow you to disaggregate a low frequency time serie with time series of higher frequency, using the French National Accounts methodology.
The aggregated sum of the resulting time-serie is strictly equal to the low-frequency serie within the benchmarking window.
Typically, the low frequency serie is an annual one, unknown for the last year, and the high frequency is either quarterly or mensual.
See "Methodology of quarterly national accounts", Insee Méthodes N°126, by Insee (2012, ISBN:978-2-11-068613-8).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/InseeFr/disaggR/issues}
}

}
\author{
\strong{Maintainer}: Arnaud Feldmann \email{arnaud.feldmann@insee.fr}

Other contributors:
\itemize{
  \item Franck Arnaud (barplot base graphics method for the mts class) [contributor]
  \item Institut national de la statistique et des études économiques \email{comptes-trimestriels@insee.fr} (https://www.insee.fr/) [copyright holder]
}

}
\keyword{internal}
