% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample.R
\name{in_sample}
\alias{in_sample}
\title{Producing the in sample predictions of a prais-lm regression}
\usage{
in_sample(object, type = "changes")
}
\arguments{
\item{x}{an object of class `praislm` or `twoStepsBenchmark`}
}
\value{
a named matrix time-serie of two columns, one for the
response and the other for the predicted value.
A `insample` class is added to the object. Then, the functions
`plot` and `autoplot` (the latter requires to load \pkg{ggplot2})
can be used to produce graphics.
}
\description{
The function \link{in_sample} returns in-sample predictions from
a \link{praislm} or a \link{twoStepsBenchmark} object.
}
\details{
The predicted values are different from the fitted values :
* they are eventually reintegrated
* the autocorrelated part of the residuals is added
Besides, changes are relative to the latest response value,
not the latest predicted value.
}
