% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.list}
\alias{model.list}
\title{Extracting all the arguments submitted to generate an object}
\usage{
model.list(object)
}
\arguments{
\item{object}{an object of class \code{"twoStepsBenchmark"}, \code{"threeRuleSmooth"}
or \code{"praislm"}.}
}
\value{
a list containing every evaluated arguments
}
\description{
The function \code{model.list} returns the arguments submitted to the function
used to generate the object of class \code{"twoStepsBenchmark"},
\code{"threeRuleSmooth"} or \code{"praislm"}.
}
\details{
These are returned as they are after evaluation, model.list doesn't
return a call.
}
\examples{
benchmark <- twoStepsBenchmark(turnover,construction); model.list(benchmark)

}
\keyword{internal}
