% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStepsBenchmark-methods.R
\name{model.list}
\alias{model.list}
\title{Extracting all the arguments submitted to generate an object}
\usage{
model.list(object)
}
\arguments{
\item{object}{a praislm or twoStepsBenchmark object.}
}
\value{
a list containing every evaluated arguments
}
\description{
The function `model.list` returns the arguments submitted
to the function \link{praislm} or a \link{twoStepsBenchmark}.
}
\details{
These are returned as they are after evaluation, model.list doesn't
return a call.
}
