% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rdir}
\alias{rdir}
\title{Random Dirichlet distrubted numbers.
Adds the gamma parameter compared to the standardt Dir variant.}
\usage{
rdir(n, alpha, gamma)
}
\arguments{
\item{gamma}{}
}
\description{
Random Dirichlet distrubted numbers.
Adds the gamma parameter compared to the standardt Dir variant.
}
