% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reView.R
\name{rePort}
\alias{rePort}
\title{Reprint an output of reView}
\usage{
rePort(object, output_file = NULL, ...)
}
\arguments{
\item{object}{either a reViewOutput object, or a character of length 1
with the path of its RDS file.}

\item{output_file}{The file in which the html should be saved. If \code{NULL}
the file is temporary, and opened in a tab of the default browser.}

\item{...}{other arguments passed to rmarkdown::render}
}
\description{
This function takes an output of the \link{reView} \pkg{shiny} application
and produces an html report with the same outputs than in shiny.
}
