% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reView.R
\name{reView_ui_module}
\alias{reView_ui_module}
\alias{reView_server_module}
\alias{reView}
\title{reView}
\usage{
reView_ui_module(id, hfserie_name, lfserie_name)

reView_server_module(id, old_bn, hfserie_name, lfserie_name, compare)

reView(
  benchmark,
  hfserie_name = deparse(benchmark$call$hfserie),
  lfserie_name = deparse(benchmark$call$lfserie),
  compare = TRUE
)
}
\description{
A shiny app to reView twoStepsBenchmarks.
}
\keyword{internal}
