% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reView.R
\name{reView_ui_module}
\alias{reView_ui_module}
\alias{reView_server_module}
\alias{reView}
\title{reView}
\usage{
reView_ui_module(id)

reView_server_module(id, old_bn, compare, function.mode = TRUE)

reView(benchmark)
}
\description{
A shiny app to reView twoStepsBenchmarks.
}
\keyword{internal}
