% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{se}
\alias{se}
\title{Extracting the standard error}
\usage{
se(object)
}
\arguments{
\item{object}{a praislm or twoStepsBenchmark object.}
}
\value{
a numeric, that is named the same way that the coefficients are.
If some coefficients are set by the user, they return \code{NA} as for
their standard error.
}
\description{
The function \code{se} returns the standard error of the coefficients
from either a \link{praislm} or a \link{twoStepsBenchmark} object.
}
\keyword{internal}
