% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praislm-methods.R
\name{se}
\alias{se}
\title{Extracting the standard error}
\usage{
se(object)
}
\arguments{
\item{object}{a praislm or twoStepsBenchmark object.}
}
\value{
a double, that is named the same way that the coefficients are.
If some coefficients are set by the user, they return `NA` as for
their standard error.
}
\description{
The function `se` returns the standard error the coefficients
from either a \link{praislm} or a \link{twoStepsBenchmark} object.
}
