% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeRuleSmooth.R
\name{threeRuleSmooth}
\alias{threeRuleSmooth}
\title{Bends a time-serie with a lower frequency one by smoothing their rate}
\usage{
threeRuleSmooth(
  hfserie,
  lfserie,
  start.benchmark = NULL,
  end.benchmark = NULL,
  start.domain = NULL,
  end.domain = NULL,
  start.mean.delta.rate = start.benchmark,
  end.mean.delta.rate = end.benchmark,
  ...
)
}
\arguments{
\item{hfserie}{the bended time-serie. It can be a matrix time-serie.}

\item{lfserie}{a time-serie whose frequency divides the frequency of \code{hfserie}.}

\item{start.benchmark}{an optional start for \code{lfserie} to bend \code{hfserie}.
Should be a double or a numeric of length 2, like a window for \code{lfserie}. If NULL, the start is defined by lfserie's window.}

\item{end.benchmark}{an optional end for \code{lfserie} to bend \code{hfserie}.
Should be a double or a numeric of length 2, like a window for \code{lfserie}. If NULL, the start is defined by lfserie's window.}

\item{start.domain}{the start of the output high-frequency serie. It also defines the smoothing window :
The low-frequency residuals will be extrapolated until they contain the smallest low-frequency window that is around the high-frequency
domain window.
Should be a double or a numeric of length 2, like a window for \code{hfserie}. If NULL, the start is defined by hfserie's window.}

\item{end.domain}{the end of the output high-frequency serie. It also defines the smoothing window :
The low-frequency residuals will be extrapolated until they contain the smallest low-frequency window that is around the high-frequency
domain window.}

\item{start.mean.delta.rate}{an optional start for the mean of the rate difference,
required for the arithmetical extrapolation.
Should be a double or a numeric of length 2, like a window for \code{lfserie}. If NULL, the start is defined by lfserie's window.}

\item{end.mean.delta.rate}{an optional end for the mean of the rate difference,
required for the arithmetical extrapolation.
Should be a double or a numeric of length 2, like a window for \code{lfserie}. If NULL, the end is defined by lfserie's window.}

\item{\dots}{if the dots contain a cl item, its value overwrites the value
of the returned call. This feature allows to build wrappers.}
}
\value{
threeRuleSmooth returns an object of class "\code{threeRuleSmooth}".

The function \code{summary} can be used to obtain and print a summary of the regression used by the benchmark.
The functions \code{plot} and \code{autoplot} (the generic from \pkg{ggplot2}) produce graphics of the benchmarked
serie and the bending serie.
The generic accessor functions \code{as.ts}, \code{model.list}, \code{smoothed.rate} extract
various useful features of the returned value.

An object of class "\code{threeRuleSmooth}" is a list containing the following components :
\item{benchmarked.serie}{a time-serie, that is the result of the benchmark.}
\item{smoothed.rate}{the smoothed rate of the threeRuleSmooth.}
\item{delta.rate}{the low-frequency delta of the rate, used to extrapolate
the rate time-serie. It is estimated as the mean value in the specified window.}
\item{model.list}{a list containing all the arguments submitted to the function.}
\item{call}{the matched call (either of twoStepsBenchmark or annualBenchmark)}
}
\description{
threeRuleSmooth bends a time-serie with a time-serie of a lower
frequency. The procedure involved is a proportional Denton benchmark.

Therefore, the resulting time-serie is the product of the high-frequency input
with a smoothed rate. This latter is extrapolated using an arithmetic sequence.

As in any disaggregation, the resulting time-serie is equal to the
low-frequency serie after aggregation.
}
\examples{

## How to use threeRuleSmooth

smooth <- threeRuleSmooth(hfserie = turnover,
                          lfserie = construction)
as.ts(smooth)
coef(smooth)
summary(smooth)
library(ggplot2)

}
