% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStepsBenchmark.R
\name{twoStepsBenchmark}
\alias{twoStepsBenchmark}
\title{Bends a time-serie with a lower frequency one}
\usage{
twoStepsBenchmark(
  hfserie,
  lfserie,
  include.differenciation = FALSE,
  include.rho = FALSE,
  set.coeff = numeric(),
  set.const = numeric(),
  start.coeff.calc = NA_real_,
  end.coeff.calc = NA_real_,
  ...
)
}
\arguments{
\item{hfserie}{the bended time-serie.}

\item{lfserie}{a time-serie whose frequency divide the higher one.}

\item{include.differenciation}{a boolean of length 1. TRUE if the regression includes a differenciation}

\item{include.rho}{a boolean of length 1. TRUE if the regression includes a rho}

\item{set.coeff}{an optional double of length 1. The coefficient of x in the regression}

\item{set.const}{an optional constant of length 1. The constant in the regression}

\item{start.coeff.calc}{the end for the coefficients estimation}
}
\value{
A time serie. Its domain is the same as the high frequency serie.
}
\description{
Bends a time-serie with a time serie of a lower frequency.
The process followed is an additive Denton, after a regression whose coefficients are
specified in regcoeffs. If one is not specified they are calculated.
}
